Refactor and correct type annotations
The YAML spec technically allows for any node kind, not just a scalar,
to be a key of a mapping node. However, in practice, this is limited by
Python requiring keys to be hashable.

Scalars have a quite general definition. Their actual values are
determined by tags, which can be arbitrary since PyYAML supports
defining new ones. Furthermore, PyYAML supports "complex Python tags",
which effectively makes it possible for scalars to be of literally any
Python type.

Therefore, when dealing with mapping nodes, it's incorrect to assume
keys will be strings, let alone any of the types corresponding to the
standard YAML tags. Even when using the SafeLoader, the latter cannot
be assumed due to support for marking arbitrary Python objects as safe
by subclassing YAMLObject and setting its loader to SafeLoader.

If using the SafeLoader, a strictly accurate type annotation for
scalars could be defined as a union of the types of the standard YAML
tags with YAMLObject. However, YAMLObject doesn't say much on its own
and therefore, in practice, a type annotation of Any is more
appropriate.

Remove the typehints module. As explained above, the annotations in
there are not accurate. Even if they were to be corrected, there isn't
much value to defining them in a separate module since they're only used
in one place. EntryType is technically fine, but having an alias for Any
doesn't seem worthwhile.

Use generic types from the typing module over collections.abc since the
latter is only supported by Python 3.9.